examples/python/custom_workflows/instruction_executor/requirements.txt
agenta
openai
pydantic
opentelemetry-instrumentation-openai
uvicorn
